Feature/pair trading float price

+++ pkg/api/bybit/mock/bybitApiMock.go
@@ -45,31 +45,31 @@ func (api *BybitApiMock) GetKlinesFutures(coin *domains.Coin, interval string, l
 	return nil, errors.New("Not implemented for Bybit API mock")
 }
 
-func (api *BybitApiMock) OpenFuturesOrder(coin *domains.Coin, amount float64, price int64, futuresType futureType.FuturesType, stopLossPriceInCents int64) (api.OrderResponseDto, error) {
+func (api *BybitApiMock) OpenFuturesOrder(coin *domains.Coin, amount float64, price float64, futuresType futureType.FuturesType, stopLossPriceInCents float64) (api.OrderResponseDto, error) {
 	return &orderResponseMockDto{
 		price:  price,
 		amount: amount,
 	}, nil
 }
-func (api *BybitApiMock) CloseFuturesOrder(coin *domains.Coin, openedTransaction *domains.Transaction, price int64) (api.OrderResponseDto, error) {
+func (api *BybitApiMock) CloseFuturesOrder(coin *domains.Coin, openedTransaction *domains.Transaction, price float64) (api.OrderResponseDto, error) {
 	return &orderResponseMockDto{
 		price:  price,
 		amount: openedTransaction.Amount,
 	}, nil
 }
 
-func (api *BybitApiMock) GetCurrentCoinPriceForFutures(coin *domains.Coin) (int64, error) {
+func (api *BybitApiMock) GetCurrentCoinPriceForFutures(coin *domains.Coin) (float64, error) {
 	return 0, errors.New("Shouldn't be called.")
 }
 
-func (api *BybitApiMock) GetCurrentCoinPrice(coin *domains.Coin) (int64, error) {
+func (api *BybitApiMock) GetCurrentCoinPrice(coin *domains.Coin) (float64, error) {
 	return 0, errors.New("Shouldn't be called.")
 }
 
 var countOfNotSoldTransactions = 0
 var maxCountOfNotSoldTransactions = 0
 
-func (api *BybitApiMock) BuyCoinByMarket(coin *domains.Coin, amount float64, price int64) (api.OrderResponseDto, error) {
+func (api *BybitApiMock) BuyCoinByMarket(coin *domains.Coin, amount float64, price float64) (api.OrderResponseDto, error) {
 	countOfNotSoldTransactions = countOfNotSoldTransactions + 1
 
 	if countOfNotSoldTransactions > maxCountOfNotSoldTransactions {
@@ -83,7 +83,7 @@ func (api *BybitApiMock) BuyCoinByMarket(coin *domains.Coin, amount float64, pri
 	}, nil
 }
 
-func (api *BybitApiMock) SellCoinByMarket(coin *domains.Coin, amount float64, price int64) (api.OrderResponseDto, error) {
+func (api *BybitApiMock) SellCoinByMarket(coin *domains.Coin, amount float64, price float64) (api.OrderResponseDto, error) {
 	countOfNotSoldTransactions = countOfNotSoldTransactions - 1
 
 	return &orderResponseMockDto{
@@ -119,20 +119,20 @@ func (api *BybitApiMock) GetActiveFuturesConditionalOrder(coin *domains.Coin, co
 }
 
 type orderResponseMockDto struct {
-	price  int64
+	price  float64
 	amount float64
 }
 
-func (d *orderResponseMockDto) CalculateAvgPrice() int64 {
-	return d.price
+func (d *orderResponseMockDto) CalculateAvgPrice() float64 {
+	return float64(d.price)
 }
 
-func (d *orderResponseMockDto) CalculateTotalCost() int64 {
-	return int64(float64(d.price) * d.amount)
+func (d *orderResponseMockDto) CalculateTotalCost() float64 {
+	return d.price * d.amount
 }
 
-func (d *orderResponseMockDto) CalculateCommissionInUsd() int64 {
-	return int64(float64(d.CalculateTotalCost()) * ((0.0001 + 0.0006) / 2)) // (0.06%+0.01%)/2    Taker Fee Rate=0.06%   Maker Fee Rate=0.01%
+func (d *orderResponseMockDto) CalculateCommissionInUsd() float64 {
+	return float64(d.CalculateTotalCost()) * ((0.0001 + 0.0006) / 2) // (0.06%+0.01%)/2    Taker Fee Rate=0.06%   Maker Fee Rate=0.01%
 }
 
 func (d *orderResponseMockDto) GetAmount() float64 {
