+++ pkg/service/exchange/exchangeDataService.go
@@ -37,13 +37,13 @@ type DataService struct {
 }
 
 //Deprecated: use GetCurrentPriceWithInterval instead
-func (s *DataService) GetCurrentPrice(coin *domains.Coin) (int64, error) {
+func (s *DataService) GetCurrentPrice(coin *domains.Coin) (float64, error) {
 	interval := viper.GetInt("strategy.trendMeter.interval")
 
 	return s.GetCurrentPriceWithInterval(coin, interval)
 }
 
-func (s *DataService) GetCurrentPriceWithInterval(coin *domains.Coin, interval int) (int64, error) {
+func (s *DataService) GetCurrentPriceWithInterval(coin *domains.Coin, interval int) (float64, error) {
 	if s.Clock.NowTime().Minute()%interval == 0 {
 		strategyIntervalString := strconv.Itoa(interval)
 		if kline, _ := s.klineRepo.FindOpenedAtMoment(coin.Id, util.RoundToMinutes(s.Clock.NowTime()), strategyIntervalString); kline != nil {
@@ -58,7 +58,7 @@ func (s *DataService) GetCurrentPriceWithInterval(coin *domains.Coin, interval i
 	return currentCoinPrice, err
 }
 
-func (s *DataService) GetCurrentPriceForFutures(coin *domains.Coin, interval int) (int64, error) {
+func (s *DataService) GetCurrentPriceForFutures(coin *domains.Coin, interval int) (float64, error) {
 	if s.Clock.NowTime().Minute()%interval == 0 {
 		strategyIntervalString := strconv.Itoa(interval)
 		if kline, _ := s.klineRepo.FindOpenedAtMoment(coin.Id, util.RoundToMinutes(s.Clock.NowTime()), strategyIntervalString); kline != nil {
